refactor(proposals): split proposal_blocks into module + define_blocks! macro
